feat: check that app is installed before installing config

+++ alacritty/install.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# check if alacritty is installed
+if ! which alacritty &> /dev/null
+then
+    echo "alacritty is not installed. Please install alacritty first."
+    exit 1
+fi
+
 # remove existing alacritty config
 rm -f $CONFIGS/alacritty/alacritty.toml
